building with nix
Miscellaneous development dependencies

add remaining dependencies for full build-test in `nix-shell --pure`.

- `ruby` (required for test scripts)
- `hlint`, `hindent` for dev

TEST(jl) guard patterns in flags, pass locally

FIX(jl): linting deps

nix-shell based building

QuickCheck ALU operation PoC

proposition module

destructuring

FIXUP(jl) hardcoded low overflow flag

FIX(jl) targets and no `stat` CI target

FIX(jl): linting pass
