+++ .github/workflows/ci.yml
@@ -22,23 +22,18 @@ jobs:
       with:
         ghc-version: '8.10.7'
         enable-stack: true
+    - uses: cachix/install-nix-action@v16
+      with:
+        nix_path: nixpkgs=channel:nixos-unstable
     - name: deps
       run: |
         sudo apt update
         sudo apt install -y iverilog verilator hlint
         iverilog -V
         verilator --version
-    - name: lint
-      run: make lint
     - name: build
-      run: make
-    - name: cache
-      uses: actions/cache@v3
-      with:
-        path: |
-          ${{ steps.setup-haskell.outputs.stack-root }}/snapshots
-        key: clash
+      run: nix-shell --run "make"
     - name: test
-      run: make check
-    - name: stat
-      run: make stat
+      run: nix-shell --run "make check"
+    - name: lint
+      run: nix-shell --run "make lint"
