+++ test/Main.hs
@@ -1,0 +1,10 @@
+{-# LANGUAGE BinaryLiterals, TemplateHaskell #-}
+
+import PropAlu
+import Test.QuickCheck
+import Control.Monad (void)
+
+return []
+
+main :: IO ()
+main = void check
